+++ udc/__init__.py
@@ -3,5 +3,5 @@ from .ignore import GitIgnore  # NOQA F401
 from .main import app  # NOQA F401
 from .types import Getable, Listable, Putable  # NOQA F401
 from .un.un_cli import UnCli  # NOQA F401
+from. un.un_uri import UnUri, K_HOST, K_PROT, K_QRY, K_TOOL, K_URI  # NOQA F401
 from .un.un_yaml import UnYaml  # NOQA F401
-from .uri import K_PROT, K_QRY, K_TOOL, UdcUri  # NOQA F401