+++ udc/benchling/entry.py
@@ -1,26 +1,24 @@
-from udc import UdcUri
-
 from .root import BenchlingRoot, BenchlingById
 
 class BenchlingEntryList(BenchlingById):
-    def __init__(self, uri: UdcUri) -> None:
-        super().__init__(uri)
+    def __init__(self, attrs: dict) -> None:
+        super().__init__(attrs)
 
     def pages(self):
         return BenchlingRoot.CLIENT.entries.list_entries()
     
 class BenchlingEntry(BenchlingRoot):
-    def __init__(self, uri: UdcUri) -> None:
-        super().__init__(uri)
+    def __init__(self, attrs: dict) -> None:
+        super().__init__(attrs)
 
     def fetch(self):
         self.entry = BenchlingRoot.CLIENT.entries.get_entry_by_id(self.id)
         self.schema = self.entry.schema.id
         self.children = {
             "authors": [author.id for author in self.entry.authors],
-            "custom_fields": [key for key in self.entry.custom_fields.additional_keys],
+            "custom_fields": [self.qoute(key) for key in self.entry.custom_fields.additional_keys],
             "days": [day.date for day in self.entry.days],
-            "fields": [key for key in self.entry.custom_fields.additional_keys],
+            "fields": [self.quote(key) for key in self.entry.fields.additional_keys],
         }
 
     def wrap(self, id, sub_type):