+++ udc/benchling/resource.py
@@ -1,6 +1,3 @@
-# URI: benchling+https://DNS#resource=entry&id=entry_id&authors=author_id
-from udc import UdcUri
-
 from .entry import BenchlingEntry, BenchlingEntryList
 from .root import BenchlingRoot
 from .sequence import BenchlingSequence, BenchlingSequenceList
@@ -22,12 +19,11 @@ RESOURCES=['aa_sequences', 'api', 'apps', 'assay_results', 'assay_runs', 'blobs'
             'schemas', 'tasks', 'teams', 'users', 'v2', 'warehouse', 'workflow_outputs',
             'workflow_task_groups', 'workflow_tasks']
 
-def BenchlingResource(uri: UdcUri):
-    root = BenchlingRoot(uri)
-    types = root.type.split('.')
-    type = types[0]
+def BenchlingResource(attrs: dict):
+    root = BenchlingRoot(attrs)
+    type = root.type
     klasses = RESOURCE_MAP.get(type)
     if not klasses:
-        raise ValueError(f"Unknown resource type[{type}]: {uri}")
+        raise ValueError(f"Unknown resource type[{type}]: {attrs}")
     klass = klasses[0] if not root.id else klasses[1]
-    return klass(uri)
+    return klass(attrs)