+++ udc/benchling/sequence.py
@@ -1,16 +1,14 @@
-from udc import UdcUri
-
 from .root import BenchlingRoot
 
 
 class BenchlingSequence(BenchlingRoot):
-    def __init__(self, uri: UdcUri) -> None:
-        super().__init__(uri)
+    def __init__(self, attrs: dict) -> None:
+        super().__init__(attrs)
 
 
 class BenchlingSequenceList(BenchlingRoot):
-    def __init__(self, uri: UdcUri) -> None:
-        super().__init__(uri)
+    def __init__(self, attrs: dict) -> None:
+        super().__init__(attrs)
 
     def pages(self):
         return BenchlingRoot.CLIENT.dna_sequences.list()