+++ udc/quilt/resource.py
@@ -1,7 +1,6 @@
 from quiltplus import QuiltResource as Resource
-from udc import UdcUri
 
 
-def QuiltResource(uri: UdcUri):
-    uri_string = str(uri)
-    return Resource(uri_string)
+def QuiltResource(attrs: dict):
+    uri = attrs.get("_uri")
+    return Resource(uri)